re-factor the code package to make navigation and searching easier
