+++ KrampitzKreutzwiserSettlersOfCatan/src/krampitzkreutzwisersettlersofcatan/gui/ClientSettings.form
