+++ KrampitzKreutzwiserSettlersOfCatan/src/krampitzkreutzwisersettlersofcatan/gui/ClientSettings.java
@@ -3,8 +3,9 @@
  * Mar 25 2021
  * Allow the user to select settings that are constant across all games.
  */
-package krampitzkreutzwisersettlersofcatan;
+package krampitzkreutzwisersettlersofcatan.gui;
 
+import krampitzkreutzwisersettlersofcatan.gui.MainMenu;
 import java.awt.Toolkit;
 import textures.ImageRef;
