+++ KrampitzKreutzwiserSettlersOfCatan/src/krampitzkreutzwisersettlersofcatan/gui/GamePanel.java
@@ -3,9 +3,19 @@
  * Nov 8, 2020
  * The JPanel that is the main part of the game. Handels drawing and logic for the game.
  */
-package krampitzkreutzwisersettlersofcatan;
+package krampitzkreutzwisersettlersofcatan.gui;
 
 // <editor-fold defaultstate="collapsed" desc="Imports"> 
+import krampitzkreutzwisersettlersofcatan.sockets.CatanClient;
+import krampitzkreutzwisersettlersofcatan.worldObjects.Port;
+import krampitzkreutzwisersettlersofcatan.worldObjects.NodeRoad;
+import krampitzkreutzwisersettlersofcatan.worldObjects.NodeSettlement;
+import krampitzkreutzwisersettlersofcatan.worldObjects.buttons.SettlerBtn;
+import krampitzkreutzwisersettlersofcatan.worldObjects.buttons.SettlerLbl;
+import krampitzkreutzwisersettlersofcatan.worldObjects.buttons.SettlerComponent;
+import krampitzkreutzwisersettlersofcatan.worldObjects.buttons.SettlerRadioBtn;
+import krampitzkreutzwisersettlersofcatan.worldObjects.Tile;
+import krampitzkreutzwisersettlersofcatan.worldObjects.WorldObject;
 import Audio.AudioRef;
 import dataFiles.OldCode;
 import java.awt.BasicStroke;
@@ -40,6 +50,8 @@ import javax.sound.sampled.UnsupportedAudioFileException;
 import javax.swing.ImageIcon;
 import javax.swing.JFileChooser;
 import javax.swing.JOptionPane;
+import krampitzkreutzwisersettlersofcatan.Catan;
+import krampitzkreutzwisersettlersofcatan.util.GlobalDataRecord;
 import textures.ImageRef;
 import static textures.ImageRef.*;
 // </editor-fold>
@@ -6445,8 +6457,8 @@ public class GamePanel extends javax.swing.JPanel implements MouseMotionListener
         }
 
         g2d.drawImage(btnImage,
-                btn.xPos,
-                btn.yPos,
+                btn.getXPos(),
+                btn.getYPos(),
                 width,
                 height, null);
     }
