+++ KrampitzKreutzwiserSettlersOfCatan/src/krampitzkreutzwisersettlersofcatan/gui/JoinOnlineGameMenu.java
@@ -3,8 +3,11 @@
  * Mar 25 2021
  * Allow the user to select settings that are constant across all games.
  */
-package krampitzkreutzwisersettlersofcatan;
+package krampitzkreutzwisersettlersofcatan.gui;
 
+import krampitzkreutzwisersettlersofcatan.sockets.CatanClient;
+import krampitzkreutzwisersettlersofcatan.sockets.CatanServer;
+import krampitzkreutzwisersettlersofcatan.gui.MainMenu;
 import textures.ImageRef;
 
 /**
