+++ KrampitzKreutzwiserSettlersOfCatan/src/krampitzkreutzwisersettlersofcatan/gui/MainMenu.java
@@ -3,14 +3,17 @@
  * Nov 06, 2020
  * The main menu for the game Settlers of Catan
  */
-package krampitzkreutzwisersettlersofcatan;
+package krampitzkreutzwisersettlersofcatan.gui;
 
+import krampitzkreutzwisersettlersofcatan.gui.NewGameSettings;
+import krampitzkreutzwisersettlersofcatan.gui.NewOnlineGameMenu;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.util.Scanner;
 import javax.swing.JFileChooser;
 import javax.swing.JOptionPane;
 import javax.swing.filechooser.FileFilter;
+import krampitzkreutzwisersettlersofcatan.Catan;
 import textures.ImageRef;
 
 /**
