+++ KrampitzKreutzwiserSettlersOfCatan/src/krampitzkreutzwisersettlersofcatan/gui/NewGameSettings.java
@@ -3,8 +3,9 @@
  * Feb 14 2021
  * Allow the user to select settings before starting a new game.
  */
-package krampitzkreutzwisersettlersofcatan;
+package krampitzkreutzwisersettlersofcatan.gui;
 
+import krampitzkreutzwisersettlersofcatan.gui.NewOnlineGameMenu;
 import textures.ImageRef;
 
 /**
