+++ KrampitzKreutzwiserSettlersOfCatan/src/krampitzkreutzwisersettlersofcatan/gui/UserManualUI.java
@@ -1,10 +1,11 @@
 /*
  * Lukas Krampitz and Evan Kreutwiser
  * Oct 23, 2020, November 4, 2020, and Nov 6 2020
- * Show the credits
+ * Show the User manual so that the usercan play the game
  */
-package krampitzkreutzwisersettlersofcatan;
+package krampitzkreutzwisersettlersofcatan.gui;
 
+import krampitzkreutzwisersettlersofcatan.gui.MainMenu;
 import dataFiles.OldCode;
 import java.io.InputStream; // File reference
 import java.util.Scanner; // File reading
@@ -15,66 +16,69 @@ import textures.ImageRef;
  * @author Tacitor
  * @author Evan
  */
-public class CreditsUI extends javax.swing.JFrame {
-    
+public class UserManualUI extends javax.swing.JFrame {
+
     private final MainMenu mainMenuFrame;
 
     /**
-     * Creates new form CreditsUI
+     * Creates new frame UserManualUI
+     *
      * @param m The main menu JFrame this returns to on exit
      */
-    public CreditsUI(MainMenu m) {
+    public UserManualUI(MainMenu m) {
         setIcon();
-        
+
         initComponents();
 
         this.setLocationRelativeTo(null);
 
         mainMenuFrame = m;
-        
-        // Load and display the credits file
+
+        // Load and display the User Manual file
         loadMaterial();
     }
 
-    /**
-     * Read and display the credits in the data file
-     */
-    public final void loadMaterial() {
-        
-        // Declare variablesD
-        Scanner fileReader;
-        InputStream file = OldCode.class.getResourceAsStream("credits.txt");
-        String fileContents = "";
-
-        // Try to read the file
-        try {
-            // Create the scanner to read the file
-            fileReader = new Scanner(file);
-        
-            // Read the entire file into a string
-            while (fileReader.hasNextLine()) {
-                // Read the line of the file into a line of the string
-                fileContents += fileReader.nextLine() + "\n";
-            }
-        }
-        catch (Exception e) {
-            // Set the sring to be displayed to an error message
-            fileContents = "Error: credits file could not be read.";
-            // Output the jsvs error to the standard output
-            System.out.println("Error reading credits file: " + e);
-        }
-        
-        // Display the file's contents from the string
-        creditslTxtAr.setText(fileContents);
-    }
-    
     /**
      * Set the icon for the JFRame
      */
     private void setIcon() {
         this.setIconImage(ImageRef.ICON);
     }
-    
+
+    /**
+     * Read and display the user manual in the data file
+     */
+    public final void loadMaterial() {
+
+        // Declare variables
+        Scanner fileReader;
+        InputStream file = OldCode.class.getResourceAsStream("userManual.txt");
+        String fileContents = "";
+
+        // Try to read the file
+        try {
+            // Create the scanner to read the file
+            fileReader = new Scanner(file);
+
+            // Read the entire file into a string
+            while (fileReader.hasNextLine()) {
+                // Read the line of the file into a line of the string
+                fileContents += fileReader.nextLine() + "\n";
+            }
+        } catch (Exception e) {
+            // Set the sring to be displayed to an error message
+            fileContents = "Error: User manual file could not be read.";
+            // Output the jsvs error to the standard output
+            System.out.println("Error reading User manual file: " + e);
+        }
+
+        // Display the file's contents from the string
+        userManualTxtAr.setText(fileContents);
+
+        // Move the invisble caret to the begining of the text to scroll to the top
+        userManualTxtAr.setCaretPosition(0);
+    }
+
     /**
      * This method is called from within the constructor to initialize the form.
      * WARNING: Do NOT modify this code. The content of this method is always
@@ -87,7 +91,7 @@ public class CreditsUI extends javax.swing.JFrame {
         backBtn = new javax.swing.JButton();
         titleLbl = new javax.swing.JLabel();
         jScrollPane1 = new javax.swing.JScrollPane();
-        creditslTxtAr = new javax.swing.JTextArea();
+        userManualTxtAr = new javax.swing.JTextArea();
 
         setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
 
@@ -98,16 +102,16 @@ public class CreditsUI extends javax.swing.JFrame {
             }
         });
 
-        titleLbl.setText("Credits:");
+        titleLbl.setText("User Manual:");
 
-        creditslTxtAr.setEditable(false);
-        creditslTxtAr.setColumns(20);
-        creditslTxtAr.setLineWrap(true);
-        creditslTxtAr.setRows(5);
-        creditslTxtAr.setText("Read carefully, this is important to play the game:");
-        creditslTxtAr.setWrapStyleWord(true);
-        creditslTxtAr.setAutoscrolls(false);
-        jScrollPane1.setViewportView(creditslTxtAr);
+        userManualTxtAr.setEditable(false);
+        userManualTxtAr.setColumns(20);
+        userManualTxtAr.setLineWrap(true);
+        userManualTxtAr.setRows(5);
+        userManualTxtAr.setText("Read carefully, this is important to play the game:");
+        userManualTxtAr.setWrapStyleWord(true);
+        userManualTxtAr.setAutoscrolls(false);
+        jScrollPane1.setViewportView(userManualTxtAr);
 
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
         getContentPane().setLayout(layout);
@@ -141,6 +145,7 @@ public class CreditsUI extends javax.swing.JFrame {
 
     /**
      * The event for when the back button is pressed. Returns to the main menu.
+     *
      * @param evt The event generated by the button click (Unused)
      */
     private void backBtnActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_backBtnActionPerformed
@@ -148,11 +153,11 @@ public class CreditsUI extends javax.swing.JFrame {
         this.setVisible(false);
         mainMenuFrame.setVisible(true);
     }//GEN-LAST:event_backBtnActionPerformed
-    
+
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private javax.swing.JButton backBtn;
-    private javax.swing.JTextArea creditslTxtAr;
     private javax.swing.JScrollPane jScrollPane1;
     private javax.swing.JLabel titleLbl;
+    private javax.swing.JTextArea userManualTxtAr;
     // End of variables declaration//GEN-END:variables
 }
