+++ KrampitzKreutzwiserSettlersOfCatan/src/krampitzkreutzwisersettlersofcatan/worldObjects/buttons/SettlerBtn.java
@@ -3,10 +3,11 @@
  * Mar 13, 2021
  * A custom button that fits the sytle of the game better and is easier to use than Swing button
  */
-package krampitzkreutzwisersettlersofcatan;
+package krampitzkreutzwisersettlersofcatan.worldObjects.buttons;
 
 import java.awt.Image;
 import javax.swing.ImageIcon;
+import krampitzkreutzwisersettlersofcatan.worldObjects.WorldObject;
 import textures.ImageRef;
 import static textures.ImageRef.ERROR_IMAGE;
